RandomHandwrittenDigitsGenerator/requirements.txt
tensorflow[and-cuda]==2.14.0
keras==2.14.0
matplotlib==3.7.5
Pillow==9.0.1
numpy==1.24.3
ipython==8.28.0
imageio==2.35.1